Optimize repeat
This pretty seriously optimizes the repeat directive.

Now, it creates an extra text node for every item (it creates an explicit start node for every item, instead of reusing the previous' end node). That's not _perfect_, but it simplifies a lot of code and makes it run _much_ faster. I think it's a pretty good tradeoff.

Fixes #95.
Fixes #157.