Switch to using text nodes with a marker instead of empty comment nodes
Comment nodes are problematic when inserted into textContent, e.g. a `<style>`
element. This commit changes to using a text node with a marker content rather
than an empty comment. The markers are removed at clone time.

All of this is implemented using a platform check to ensure that we only do
this on platforms that don't clone empty text nodes (IE).

Fixes #214