Hoist template preparers to avoid excessive closures
Instead of recreating the `_prepareTemplate` and `_prepareInstance`
helpers on every template creation, this hoists them to prototype
functions and passes in the necessary params. It's a bit verbose, but it
should help.

Also, it avoids re-creating a `TreeWalker` on in every nested
`<template>`. Any constructor that touches WebIDL is dog slow. Reuse the
same instance all the way down the recursion.