Enable TypeScript declaration maps (#528)
Declaration maps are a neat feature that landed in TS 2.9, enabling it
emits a sourcemap for declaration files (.d.ts.map). Like JS .map files
they are for resolving to the original source of declaration files, but
instead of being used by the browser they are used by the user's text
editor/ide.

Basically, where before in vscode if you go to defintion on `html` in
`import {html} from 'lit-html'` you would end up in the declaration
file, now it will take you to the correct position in the original .ts file.