Fix directive syntax in docs
Seems like the documentation has incorrect examples on how to define a directive.

I tried to do it according to the docs and saw the stringified function in the rendered DOM. Then I looked the implementation of existing ones in the core and made mine in the same way (https://codepen.io/bashmish/pen/oQxBYZ) and it started working. I fixed examples accordingly.