initial impl for removal notification for repeat directive
A very "naive" implementation of `onRemove` just to get the ball rolling. It currently has the following limitations:

- API shape has be thought through.
- `markerNodes` are also notified of removed, which is probably is better filtered off.
- Should this just be a part of repeat, or fit it in a larger scale?

Please let me know if fits into your direction for the project. If it does, we can start ironing out the details, and work on this commit. That being said, it works as expected. 

Related issue: #635