Cleanup Template Preperation (#696)
By storing the `endNode`'s index (instead of the `startNode`), it becomes easier to reason about the `COMMENT_NODE` path.

The only reason it's complicated now is because we attempt to remove the node, provided there's some `nextSibling`. If we removed that, `COMMENT_NODE` would become dead simple (at the expensive of a unnecessary extra node).