Cache only on the TemplateStringsArray
This is a refactoring of #579.

The `keyString` caching from #579 permantly caches every `Template`, which holds its `<template>`. This isn't ideal. So, we can't use the `TemplateStringsArray` string values as a caching key.

The issue here is that we want to prevent expensive computations as much as possible. With rendering (and re-rendering when changing to a new template), there are two big computations:

1. `<template>` creation and finding `TemplatePart`s
2. Thrashing a DOM tree

`#1` is due to creating a new `Template`. `#2` is due to creating a `TemplateInstance` from that `Template` and committing it to a container, which could have this exact DOM tree already rendered.

The current code uses the identity equality of `TemplateStringsArray` to get the same `Template` every time, preventing `#1`. It then uses the identity equality of the `Template` instance to avoid a `#2`'s thrash.

- - -

This PR works around these issues by deferring the heavy computations until as late as possible.

First, I've made `Template` instances cheap to create. They now hold onto the `TemplateResult`, and do no further work until necessary. This allows me to only cache based on the `TemplateStringsArray`. Even if Safari forgets to properly cache it, creating a new `Template` doesn't hurt much.

Second, I've changed the re-render check to use deep equality of the `Template`. Because the `Template` now holds onto the `TemplateStringsArray`, I can just grab those and check if they're identity equal (good browsers) or string value equal (Safari).

If we determine the two `Template`s are equal, we can prevent the trash just like normal. If they aren't equal, we then do the two heavy computations.

This leads to an almost ideal caching situation. Unfortunately:

- a `TemplateInstance` holds the `Template`
- a `Template` holds the `TemplateResult`
- a `TemplateResult` holds the `TemplateStringsArray`
- a `TemplateStringsArray` weakly holds the `Template` (via the `WeakMap`)

So we still have a retain cycle. But, once a `NodePart` rerenders with any other value, that cycle is broken. That's not too bad, and it's the best we can do.