Cleanup until directive
Gets rid of a few useless if checks, and uses an infinity value so we don't have to explicitly check for `undefined`. Also guards against an out-of-bound array access on `previousValues`.

I took a bit of liberty with using a 2<sup>31</sup>-1 as "infinity". If you want, we can use the real `Infinity` value.