Rework TemplateInstance cloning
Instead of traversing over every single `Element`, `Comment`, and `Text`
node, this looks only for comment nodes that mark part positions.
Additionally, I'm using a comment node to mark the spot after a template
element, so that we know to go back and traverse the template.

I can anywhere between a 2-10x speedup with this:
https://jsbench.github.io/#d26f5d42b64a4c21dda947adcaaa1570