Remove unnecessary type assertions
This is the result of turning on the no-unnecessary-type-assertion tslint pass and running `--fix`, then fixing up a few tests using `Deferred` and adding an additional assertion in places where TypeScript couldn't infer that a variable was always assigned before being read.

I like this lint pass because it makes the code a bit easier to read and understand. I will admit though that I'm proposing this change because we use this lint pass downstream, and it's annoying seeing the lint warnings.