Suite functions are still run when suites are skipped. (...)
When a suite is skipped with `suite.skip`, the function passed to `suite.skip`
is still run to determine what tests and sub-suites are in it, even though none
of the contained tests are run. In this suite that depends on the global
`customElements` object, the actual use of that object needs to be wrapped in
`suiteSetup` (or a similar hook) so that it is only run when the suite's tests
are run.