WIP: Hydration experiment
This prototypes a hydration method where we:
 - differentiate opening and closing marker nodes so we can build a tree of markers from DOM
 - calculate a template hash from template strings
 - record the template hash in NodePart opening markers when rendering nested templates
 - add a hydrate() analog to render() that doesn’t clear the container

Seems to work for an extremely basic example. We would need to add attribute part support, likely by encoding attribute parts in comments that are first-children of the element that has the part. The extra markers are a cost, so we’d want to emit them only when pre-rendering.