Remove hard-coded language codes/names for now
For now, just let language codes pass through to translation formats as-is, with
just basic validation that they broadly match the alpha-numeric syntax of BCP
47.

BCP 47, or at least subsets of it like ISO 639-1, seem fairly ubiquitous, so
it's likely we could safely assume use of BCP 47 and validate more strictly with
known codes.

But BCP 47 is comprised of many specs, and is a little complicated, so I'm
not confident yet I fully understand it enough to say that we should use it
to validate. Plus locale codes can be added over time, so we'd have to keep
up to date on that.

I'll do some more research sometime, and possibly integrate libraries like
https://github.com/wooorm/bcp-47 and https://github.com/adlawson/nodejs-langs.