Support customization of locale initialization
Previously, the locale was always initialized by looking for the "locale"
URL parameter. This behavior was inherited from Pathfinder.

Now the following options are available:

- Export a setLocale function. Users can use this to set locale however they
  want.

- Initialize locale from a regexp, or a configurable URL parameter.

Note that the advantage of initializing locale directly from the localization
module is that it means there is no possibility of the locale being stale, since
initialization must occur before the first msg call is made.

The config format is also refactored to allow for additional output modes.
Things I'm thinking of for upcoming work:

- Rewrite mode, which rewrites the whole program for each locale (I'm planning
  to do this one next).

- A way to trigger re-renders whenever the locale changes. I was thinking of
  a mixin for LitElement that would keep track of instances and force a render
  on all of them whenever setLocale is called.

- A way to fetch messages dynamically, instead of all of them being hard-bundled
  into the generated module. In this case, maybe setLocale would return a
  promise, and the above mixin would know to wait for them.