[lit-next] shady render (#1315)
Adds `platform-support` module to LitElement. This replaces the previous `shady-render` module in lit-html and provides `ShadyCSS` style scoping support for both `lit-html` and `LitElement`. 

To use, (1) load the web components polyfills, (2) load `platform-support` (this can be loaded as a legacy script or imported as a module, (3) load Lit code.

In addition to being opt-in in LitElement, there are some other changes v. the previous `shady-render`:
* Inside `<style>` elements, Lit parts are no longer supported. They will be supported in the future via static parts.
* Unlike `shady-render` where a separate `render` method was used, `platform-support` augments Lit's base functionality to support ShadyCSS scoping.
* In `lit-html` the `ShadyCSS.styleElement` method is no longer called. For standalone use in lit-html, it must be manually called after rendering and whenever dynamic changes are made. It is still handled in LitElement.