Port sanitization hooks to lit-next.
See https://github.com/Polymer/lit-html/pull/750 for details of this API.

This is an underoptimized basic implementation. More of it can be removed from the release build, and we're not yet caching sanitizers.

I did modify the test infra so that we can test that we're caching sanitizers though.