Make `_state` an instance property of `until`. (...)
`_state` maps parts to the last set of values provided to a directive and the
index of the last value from that list that was rendered. It needs to be an
instance property because the mapping between parts and directives is not always
one-to-one. Particularly, for interpolated AttributeParts, the mapping can be
one-to-many, which can cause the multiple directives associated with that single
part to overwrite each other's state if they key only by the associated part.