[lit-next] Adds AsyncTask controller (#1489)
A controller that performs an asynchronous task like a fetch when its host element updates. The controller performs an update on the host element when the task becomes pending and when it completes. The task function must be supplied and can take a list of dependencies specified as a function that returns a list of values. The `value` property reports the completed value, and the `error` property an error state if one occurs. The `status` property can be checked for status and is of type `TaskStatus` which has states for initial, pending, complete, and error. The `render` method accepts an object with optional corresponding state method to easily render values corresponding to the task state.