Optionally enables support for `ShadyDOM.noPatch` setting (#1525)
Adds ShadyDOM.noPatch support and tests to ReactiveElement, LitElement, and lit-html. Support is on by default in dev mode and must be opted into in production mode by setting `ENABLE_SHADYDOM_NOPATCH`. In noPatch mode, ShadowDOM api is patched directly to ReactiveElement prototype which avoids the need to wrap ReactiveElements.