[localize] Fix duplicate and out-of-order localized template values (#1898)
Fixes two bugs relating to localized template values:

When we generated localized templates, we replace expressions with numbers, representing the position of that value in the source template, to account for the possibility of expressions changing position (see #1625 for more details). However, we were incorrectly indexing using the position of the <ph> placeholder that contained the expression, instead of the index of the expression itself (note that a <ph> can contain any combination of HTML and zero or more placehlders). So the offsets could be wrong, and were sometimes duplicated.

We use the special string _START_LIT_LOCALIZE_EXPR_${RANDOM}_ as a delimiter during internal template decomposition, but a replace operation was accidentally only matching the first occurrence, instead of all occurrences. So a template like <a href="${url}/${base}"> previously contained this literal string.

Fixes #1896