fix(ifDefined): treat type as lit-analyzer would
Lit analyzer would see ifDefined function name and simply calculate the type as if it were the value without `undefined`. This fixes an issue where passing in a `DirectiveResult|undefined` was being calculated instead as `DirectiveResult | typeof nothing`. This meant lit-analyzer could not understand that this was a directive and meant to be ignored by the type checker.