@query and @queryAll return null or [] before first update (#2065)
Fixes #2062. To match Lit1 behavior, the `@query` decorator returns `null` (rather than `undefined`) if a decorated property is accessed before first update. When the `cache` setting is true, the first result is always cached and it is either `null` or the queried element. Likewise, a `@queryAll` decorated property returns `[]` rather than `undefined`.