[ts-transformers] Preserve blank lines and fix constructor ordering (#2067)
Adds two new transformers to the @lit/ts-transformers package: preserveBlankLinesTransformer and constructorCleanupTransformer.

The goal of these transformers is to address the loss of formatting that occurs in any standard TypeScript compile. We care about preserve formatting because we're using these transforms to generate idiomatic JavaScript from TypeScript that we will present in lit.dev documentation (lit/lit.dev#332).

The problems these address are:

TypeScript never emits blank newlines. The preserveBlankLinesTransformer transformers addresses this by inserting //__BLANK_LINE_PLACEHOLDER_G1JVXUEBNCL6YN5NFE13MD1PT3H9OIHB__ comments wherever a blank line appeared in the original source. These can comments can then be trivially replaced with newlines as a post-transform step.

TypeScript often moves the constructor to the top of the class, above static members, even when the constructor was present in the original source code. The constructorCleanupTransformer addresses this by checking the original source position of constructors, and moving them back to their original position. If the constructor is completely synthetic (meaning it doesn't have an original source position), we do a best effort by moving it below the final static member of the class, and inserting a blank-line placeholder comment.

Also includes a bugfix related to import statement order.