[ts-transformers] Clean up unnecessary `super(...arguments)` calls (#2109)
When synthesizing a subclass constructor, TypeScript always includes a super(...arguments) call.

However, it is often unnecessary to include the ...arguments argument.

This PR updates the constructor-cleanup transformer to remove the ...arguments argument when none of the ancestor classes have any parameters in their constructor.

Note this is not 100% safe, since an ancestor constructor could directly use arguments in its body (which we may not be able to detect because we don't have access to bodies via .d.ts files) but that should be fairly rare. For the main use case we're targeting (generating lit.dev samples), it seems like a reasonable readability/correctness tradeoff.