Increase tests start timeout.
In sauce ie11 we need to give enough time for all the tests to be
run sequentially. What appears to happen with a short testsStartTimeout
is that all tests get queued up initially. Then the test runner processes
them one at a time. However the testsStartTimeout is measured from queue
time. Therefore when the last test is dequeued, if testsStartTimeout has
expired we immediately fail.

Solution is to increase start timeout to provide more time for the total
sequential tests. This is a stopgap because as we add more tests we will
need to keep raising this limit.

I'm also not completely sure if this will work and need to test this
hypothesis through Github Actions.