Use new lit-robot PAT for automating adding issues to Lit project (#2416)
I've created a new https://github.com/lit-robot account and added a personal access token (PAT) to the Lit org's secrets. We can use this for other general purpose automation too if needed.

This is required for the action that automatically adds newly filed issues to the Lit project board because the default GITHUB_TOKEN token only has permission for this repo, not for the whole org, but our project board is org-scoped.