[labs] Initial commit of eleventy plugin (#1755)
A plugin for Eleventy for pre-rendering pages that include Lit web components. The components can then be (optionally) hydrated on the client for interactivity.

Note, this plugin is primarily focused on rendering standalone widgets which can be configured with attributes in HTML into an 11ty site (there currently no facility for passing 11ty data into components via Javascript).