Work around a `changesets/action` bug with its `cwd` option. (...)
`changesets/action@v1` has a bug where it attempts to set the git user info
locally before changing to the directory specified with `cwd`.[^1] This means
that it will fail at this step if the root directory is not a git repo. To work
around this, this PR initializes a new git repo in the root directory and sets
Lit Robot's git user info globally.

[^1]: https://github.com/changesets/action/blob/898d125cee6ba00c6a11b6cadca512752c6c910c/src/index.ts#L19-L34