Add a devlog events system (#2401)
* Add a devlog events system

When running Lit in dev mode, and when window.emitLitDebugLogEvents is true, we will emit 'lit-debug' events on window. These events will be fired during rendering and as part of the ReactiveElement update lifecycle, giving internal information about rendering that may be useful for debug tools and visualizers.

Production builds will not be affected, all of the devlog code is compiled out when DEV_MODE is false.

Note that this is an unstable API which may change or even be removed in the future, even on patch releases.