[Infra] Convert SSR packages to Wireit (#2857)
Converts `ssr`, `ssr-client`, and `eleventy-plugin-lit` to Wireit.

Also changes the core library `build` scripts to include their dependencies `build` scripts. This is useful because the `:rollup` scripts don't include their dependencies `:rollup` targets (which is good, because they can build in parallel), but you often need a target to "the fully built package so I can actually run it in a browser". So that's what the `build` scripts now are, and tests will typically depend on those.

Part of https://github.com/lit/lit/issues/2724