[ssr] Fix hydration bug with attribute bindings on void elements (#2952)
### Background

In SSR, when an element has an attribute binding, we insert a `<!--lit-node 0-->` comment to tell `hydrate()` about the attribute binding.

However, if that element was a *void* element, such as `<input>` or anything from [this](https://html.spec.whatwg.org/multipage/syntax.html#void-elements) list, then when the HTML parser encounters it, all of the children it has will be hoisted up as siblings.

For example:

```ts
render() {
  return html`<input max=${this.maxLen}>`;
}
```

Is SSR'd as:

```html
<input max="42"><!--lit-node 0--></input>
```

But parsed by the browser as:

```html
<input max="42" />
<!--lit-node 0-->
```

This means when `hydrate()` encounters a case like this, we need to check `previousSibling` instead of `parentElement`, to find the element to receive the attribute binding.

### Bug

We already accounted for void elements where we actually create the attribute parts: https://github.com/lit/lit/blob/ac356997351874706f8be235559c765861dce67d/packages/lit-html/src/experimental-hydrate.ts#L340

But we did not account for it where we remove the `defer-hydration` attribute: https://github.com/lit/lit/blob/ac356997351874706f8be235559c765861dce67d/packages/lit-html/src/experimental-hydrate.ts#L161

This actually only crashed `hydrate()` if the void element was an *immediate child of a shadow root*, because in every other case, `element.parentElement` would return *something*, even if it wasn't the correct node.

### Fix

We now also account for void elements when we remove the `defer-hydration` attribute in the same way that we already did for creating attribute parts.

Note as part of this I switched from `previousSibling` to `previousElementSibling`, so that we can assume `removeAttribute` is defined. Seems like this should be safe?

Hopefully fixes https://github.com/lit/lit/issues/2946

cc @daKmoR 

### Alternative idea

I also thought about an alternative way to handle void elements generally, which is to detect them up-front during SSR rendering, since they are a small fixed set of tag names, so that we get an explicit signal about whether we need to check the parent or the sibling during hydration, instead of doing the previousSibling -> parentElement fallback. See https://github.com/lit/lit/commit/66b702b4fc976f3eaa6909e9b79a132b333f4f4e for how that would look. This could potentially have slightly better performance during hydration, since it replaces a DOM call with a string check. Not sure it's worth it though, and I think it would be a breaking change because it changes the rendering scheme.