[labs/react] useController: handle strict mode (#2987)
Fixes #2950. React expects `useState` not to have side effects. `useController` calls "connect" and "update" before returning. This ensures the value in the rendering component is as expected. However, to minimize the chance of side effects, when the controller is created, it schedules a disconnection. This disconnection is cancelled if the component effects actually run. This handles React's strict mode behavior. Also added a test that uses React.StrictMode to verify this behavior.