[ssr] Cast setAttribute values to string in SSR DOM shim (#3045)
Previously, when our SSR DOM shim for `setAttribute` received a value, it just assumed it was a string and assigned it to its internal attribute map as-is. Later, when trying to serialize an attribute to HTML, we pass it through an escape function, which crashes if the value is not a string.

In browsers, the value passed to `setAttribute` is always cast to a string. So `42` becomes `"42"`. We now mirror this behavior.

This specifically came up in https://github.com/lit/lit/issues/2633 because in `ReactiveElement`, when a property has `{type: Number}`, we do not cast the value to a string ourselves, since we assume the browser will do that automatically (in fact, we don't handle `Number` at all, since the default behavior is correct https://github.com/lit/lit/blob/21ff05104727d088ab577f7178cbc15eed14841a/packages/reactive-element/src/reactive-element.ts#L321).

Fixes https://github.com/lit/lit/issues/2633

cc @daKmoR