Relax the typescript version for localize-tools (#3800)
Fixes #3739

By relaxing the typescript version from ~4.7.4 to ^4.7.4, we allow it to use 4.8 and 4.9 as well. This fixes the problem with localize-tools where it uses typescript 4.7, but the project used 4.8 or 4.9. This breaks because of breaking changes between 4.7 and greater.

Note: so 5.0 is not valid here. I'm not sure whether we want to allow that yet.