[lit-html] fix type of isTemplateResult and then fix cache and hydration
This commit fixes the type of `isTemplateResult` to return the correct
type as the current implementation: a `TemplateResult` and a
`CompiledTemplateResult`.

Then I fixed the cache directive to work with both template types, and
hydration to only work with uncompiled templates.

Added a unit test for the compiled template case with the cache directive.

An alternative I considered:
 - Make `isTemplateResult` stricter such that the type is only a
   `TemplateResult`. That is technically a breaking change (although
   people should not be depending on `CompiledTemplateResult`s)
 - And add the following helpers: `isCompiledTemplateResult` and
   `getTemplateStringsArray`.