Support @provide with experimental decorators on a property (#4200)
* Support @provide with experimental decorators on a property

It might feel like this is working at cross purposes to our standard decorators work, because those all require accessors, but I don't think so. With standard decorators we can use the type system to error out if they're used on a property instead of an accessor. With experimental decorators, they sorta work with properties, in that they do provide values and update them, but reading the value back off the field would always give undefined because we were defining it without a getter.

* Address review comments

* Fix import.