Avoid using dynamic property access
Also hide the static property initialization from jscompiler, and it won't manually copy those fields down onto subclasses.

This is a bit larger for terser, but we expect the code to be faster because we're doing fewer dynamic property accesses, and it's both smaller and faster for jscompiler because it can rename both elementProperties and finalize.