First implementation of the toolbar. (#38)
* First implementation of the toolbar.

Can select between a select mode and an interact mode. With select mode, you get bounding boxes and with interact mode the glass goes away.

* Don't overlap toolbar with content.

* Move ignition-toolbar to its own file.

* Use @vscode/webview-ui-toolkit button

* Bundle webview and iframe code with esbuild.

Faster builds, faster page loads, what's not to love?

* CSS tweak.

* Some more style improvements.

* CSS ordering: more specific comes after

* Remove broken scroll-compensation code.

Turns out both elementsFromPoint and DOMRect are both viewport based, not page based? In any case, removing this makes it work.

* Minor fixups.

* Integrate esbuild more properly.

We now build with both TSC and esbuild, but we serve out bundled but not minified sources out of the uiserver

This improves startup time a lot, and simplifies the uiserver too, since it's now just a simple file server.

* Row layout instead of column

Co-authored-by: Justin Fagnani <justin@fagnani.com>

* Horizontal toolbar layout

Also lots of little fixes to styles.

---------

Co-authored-by: Justin Fagnani <justin@fagnani.com>