Use an updatable analyzer, with an overlay in memory filesystem.
It's kinda tricky to notice this working in practice, because we'll need to either do an HMR or reload the story frame, but the server will now update, and the analyzer should always give the latest info when queried.

To test this, you can have a couple of ignition editors going such that swapping between them will reload the stories.