[labs/virtualizer] improve type and supporting documentation
First, this fixes the `RenderItemFunction` type so it can return
any stringable value. Just as lit expects internally for rendering.

Second, the documentation in the readme is updated to reflect the
current type definition. As well as adding extra guidance on using
the generic to satisfy more strict TypeScript configurations.

Fixes: #4377