[lit-labs/virtualizer] add keyFunction docs and update jsdocs
This provides documentation in the readme for the keyFunction
property. It also adds jsdocs for other properties to provide
some inline support in editors when using the class reference.
By implementing the directive's configuration interface, the
element inherits the JSDocs from it as well. Helping us reduce
duplication across the code-base.

Also apply an `@deprecated` tag to `scrollToIndex` so it is
exposed as deprecated.